Refactor: PlanService interface 분리
