+++ src/main/java/com/trablock/web/controller/plan/DirectoryController.java
@@ -12,6 +12,7 @@ import com.trablock.web.entity.plan.Plan;
 import com.trablock.web.entity.plan.UserDirectory;
 import com.trablock.web.service.plan.PlanItemService;
 import com.trablock.web.service.plan.PlanService;
+import com.trablock.web.service.plan.PlanServiceImpl;
 import com.trablock.web.service.plan.UserDirectoryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
