+++ src/main/java/com/trablock/web/controller/plan/PlanController.java
@@ -1,6 +1,5 @@
 package com.trablock.web.controller.plan;
 
-import com.sun.xml.bind.v2.TODO;
 import com.trablock.web.controller.form.Form;
 import com.trablock.web.converter.Converter;
 import com.trablock.web.dto.plan.*;
